A C program that prints the various types using 'printf'

+++ 0x00-hello_world/6-size.c
@@ -20,7 +20,7 @@ int main(void)
 	printf("The size of a char is: %lu.\n", (unsigned long)sizeof(c));
 	printf("The size of a char is: %lu.\n", (unsigned long)sizeof(i));
 	printf("The size of a float is: %f.\n", (unsigned long)sizeof(li));
-	printf("The size of a long int is: %ld.\n",(unsigned long)sizeof(lli));
-	printf("The size of a long long int is: %lld.\n", (unsigned long)sizeof(f);
+	printf("The size of a long int is: %ld.\n",(long long int)sizeof(lli));
+	printf("The size of a long long int is: %lld.\n", (double)sizeof(f));
 	return (0);
 }
